requirements.txt
# requirements.txt

# Core dependencies
torch
torchvision
einops==0.3.2
tqdm==4.64.1
numpy==1.23.5
scikit-image==0.19.3
lpips==0.1.4
pyyaml==6.0
Pillow==9.4.0
lpips
scikit-image

# Additional dependencies
matplotlib==3.6.2
opencv-python==4.6.0.66
